+++ R/read.R
@@ -61,7 +61,7 @@ set_utf8 = function(x) {
 #'   and if necessary corrected (when seen from above: exterior ring counter
 #'   clockwise, holes clockwise)
 #' @details for \code{geometry_column}, see also
-#' \url{https://trac.osgeo.org/gdal/wiki/rfc41_multiple_geometry_fields}
+#' \url{https://gdal.org/en/latest/development/rfc/rfc41_multiple_geometry_fields.html}
 #'
 #' for values for \code{type} see
 #' \url{https://en.wikipedia.org/wiki/Well-known_text#Well-known_binary}, but
@@ -294,8 +294,6 @@ st_read.character = function(dsn, layer, ..., query = NA, options = NULL, quiet 
 
 	if (length(promote_to_multi) > 1)
 		stop("`promote_to_multi' should have length one, and applies to all geometry columns")
-
-
 
 	if (use_stream) {
 		stream = nanoarrow::nanoarrow_allocate_array_stream()